refractor: rename "controls" folder "sidebar" and move Sidebar.js to subfolder
