+++ src/components/sidebar/ColorMapControl.js
