+++ src/components/sidebar/OpacityControl.js
